Code Cleanup
Moving shared code in conftest.py
